Refactor(analysis): split package to performance-analysis
