+++ palladio-analysis/pom.xml
@@ -1,4 +1,5 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
@@ -24,6 +25,10 @@
         <dependency>
             <groupId>de.fabiankeller.palladio-headless</groupId>
             <artifactId>palladio-bridge</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>de.fabiankeller.palladio-headless</groupId>
+            <artifactId>performance-analysis</artifactId>
         </dependency>
 
         <!-- test dependencies -->
