+++ palladio-analysis/src/main/java/de/fabiankeller/palladio/RunLqnsWithBuilder.java
@@ -3,8 +3,8 @@ package de.fabiankeller.palladio;
 import de.fabiankeller.palladio.analysis.pcm2lqn.runner.PcmLqnsAnalyzer;
 import de.fabiankeller.palladio.analysis.pcm2lqn.runner.PcmLqnsAnalyzerContext;
 import de.fabiankeller.palladio.analysis.provider.SimpleTacticsProvider;
-import de.fabiankeller.palladio.analysis.result.PerformanceResult;
-import de.fabiankeller.palladio.analysis.result.Result;
+import de.fabiankeller.performance.analysis.result.PerformanceResult;
+import de.fabiankeller.performance.analysis.result.Result;
 import de.fabiankeller.palladio.builder.PcmBuilder;
 import de.fabiankeller.palladio.environment.PalladioEclipseEnvironment;
 import org.palladiosimulator.pcm.core.entity.NamedElement;
