+++ palladio-analysis/src/main/java/de/fabiankeller/palladio/analysis/pcm2lqn/results/Pcm2LqnResult.java
@@ -1,6 +1,6 @@
 package de.fabiankeller.palladio.analysis.pcm2lqn.results;
 
-import de.fabiankeller.palladio.analysis.result.AbstractPerformanceResult;
+import de.fabiankeller.performance.analysis.result.AbstractPerformanceResult;
 import org.palladiosimulator.pcm.core.entity.NamedElement;
 import org.palladiosimulator.solver.models.PCMInstance;
