+++ palladio-analysis/src/main/java/de/fabiankeller/palladio/analysis/pcm2lqn/results/Pcm2LqnResultsParser.java
@@ -1,8 +1,8 @@
 package de.fabiankeller.palladio.analysis.pcm2lqn.results;
 
-import de.fabiankeller.palladio.analysis.result.PerformanceResultWriter;
-import de.fabiankeller.palladio.analysis.result.exception.InvalidResultException;
-import de.fabiankeller.palladio.analysis.result.valueobject.Duration;
+import de.fabiankeller.performance.analysis.result.PerformanceResultWriter;
+import de.fabiankeller.performance.analysis.result.exception.InvalidResultException;
+import de.fabiankeller.performance.analysis.result.valueobject.Duration;
 import de.fabiankeller.palladio.analysis.tracing.PcmModelTrace;
 import org.palladiosimulator.pcm.core.entity.NamedElement;
 import org.palladiosimulator.pcm.resourceenvironment.ResourceContainer;
