--- palladio-analysis/src/main/java/de/fabiankeller/palladio/analysis/result/trace/TraceMapper.java
@@ -1,14 +1,0 @@
-package de.fabiankeller.palladio.analysis.result.trace;
-
-import de.fabiankeller.palladio.analysis.result.PerformanceResult;
-
-/**
- * Interface used to trace performance results between different models.
- */
-public interface TraceMapper<FROM, TO> {
-    /**
-     * Maps a performance result to a different model by mapping the elements the results are attached to from one model
-     * to the other.
-     */
-    PerformanceResult<TO> map(PerformanceResult<FROM> result);
-}
