--- palladio-analysis/src/main/java/de/fabiankeller/palladio/analysis/result/valueobject/ValueObject.java
@@ -1,12 +1,0 @@
-package de.fabiankeller.palladio.analysis.result.valueobject;
-
-/**
- * Defines value objects for the results.
- */
-public interface ValueObject {
-
-    /**
-     * Turns the value object into a human readable string.
-     */
-    String toHumanReadable();
-}
