+++ performance-analysis/src/main/java/de/fabiankeller/performance/analysis/result/exception/InvalidResultException.java
