+++ performance-analysis/src/main/java/de/fabiankeller/performance/analysis/result/metric/PerformanceMetric.java
