+++ performance-analysis/src/main/java/de/fabiankeller/performance/analysis/result/valueobject/ValueObject.java
@@ -1,0 +1,12 @@
+package de.fabiankeller.performance.analysis.result.valueobject;
+
+/**
+ * Defines value objects for the results.
+ */
+public interface ValueObject {
+
+    /**
+     * Turns the value object into a human readable string.
+     */
+    String toHumanReadable();
+}
